AllClass.R docs now in markdown format
This is partly a test, to learn the best way to phrase things. For
example, in Rd I used to write \code{\link{ctd-class}} but now when I do
the new [ctd-class], only the word "ctd" appears, and therefore I need
to add the suffix -class after the instance. I like the fact that the
link comes out just as "ctd", because that *is* the class type.

The advantage of using markdown is two-fold:
1. it's easier to write and to maintain
2. its output ought to be in a format that readers expect. Some of this
is likely subtle, e.g. the "-class" issue mentioned above.

My intention is to change over to markdown docs whenever I edit a
function. Changing a whole .R file over is a bigger thing, though,
requiring perhaps half an hour in the editor but the interwoven test
builds consume quite a lot of more time, yielding a conservative
estimate of perhaps 1h/kloc, i.e. 1 to 2h per .R file.