windrose doc: fix two broken links
OK, now I know how to do that right: you have to Rd format, e.g.

\code{\link{[[<-,windrose-method}}

I suspect roxygen may handle such things in markdown, but for now this
is fine. (The only trick is that I have to undo the action of my vim
macro that handles such things.)