remove code diagnostics about using "<-" in an arglist
I have done this for all but oce.as.POSIXlt(), since the latter is a bit
tricky and I need more time to rephrase its logic. It is a bit mangled
to me at first glance, e.g. what the heck am I doing with the
`is.na(xx)` tests ... if `xx` is `NA` then surely the return value ought
to be`NA`, so why am I doing the `||` operation?