add xbt support
See
https://github.com/dankelley/oce/issues/1494
for discussion.

I did not add [as.xbt()].  Although that would be reasonably easy, I
think it should be done only after we have a firmer idea of the data
that might be contained in arbitrary xbt objects, as opposed to sippican
ones. The problem is that if we need to make adjustments to a limited
[as.xbt()], we might need to change the arglist, and that can turn into
a nightmare for compability with code that users might have written in
the meantime.