xbt[[ improvements
Now, the object can store 'depth' if that is the name in the data file,
or 'z' if that is what is present. But [[ will retrieve 'z', 'depth' or
'pressure', by converting.  Similarly, [["soundspeed"]] recovers sound
speed by looking it up in @data (if it's there) or by calculation,
assuming S0=35.

The plot scheme is also changed to use z on the vertical axis, instead
of depth