silence a ncdf4 warning
I was getting a warning on my home machine but not my work one.  My home
is running R-devel and the work is R-release.  I do not know why they
differ, and have no easy way to explore this effectively.  Note that the
OS is also different. In any case, the main thing is that I want the
test suite to work the same for both cases, so I moved diagnostic
warning() to an oceDebug() message, so now the two systems ought to
behave identically and so now the test suite handles both.