recreate data(met) to avoid serialization
We don't need serialization for this, so there's no benefit. The problem
with serialization is that it limits users to the most recent R.
Perhaps I'll revisit this -- and rebuild a lot of datasets -- after two
more R releases.

NOTE: the changes to the R file are odd. I had thought that my tests on
the R version made sense, but now I see that they would only work for R
3.x.x, but not for the upcoming 4.x.x version.  But I do not plan to fix
up the create code because I don't want serialization, anyway.  Maybe
I'll revisit this sometime (for all oce objects) if I learn that some
oce objects can be saved more compactly with serialization.