try to fix formatting error
It seems as though a tabular environment cannot contain items consisting
of a dash.  Well, maybe that was the problem -- I'm not really sure.
However, this seems to build locally, i.e. when I do

    cd man
    R CMD Rd2pdf oce-deprecated.Rd

it works without a latex error now. But that's on a preview of the
upcoming OSX release, so to be sure I'll need to wait for 35 min until
the travisCI build is done.