tidem() improvements for issue 1621
See https://github.com/dankelley/oce/issues/1621 and note:

1. The renaming from T-TIDE was wrong (reversed) for "UPS1".

2. Now, we have a new user-visible function named
tidemConstituentNameFix() that does renaming. This will make it easier
for users to debug problems, and also submit patches (if more naming
differences are uncovered).

3. The internal functioning of Z0 is changed. The tests in
tests/testthat/test_tidem.R still work properly, which means that we are
still reproducing the T_TIDE results (to the extent that we check them).
Also, those tests require that amplitude and phase match values from
previous versions of oce, so we know that this is okay also.
Finally, there are tests that predict() works sensibly in that file.
Given all of this, I am fairly confident that things are OK, at least
for tidem() and predict().  (I have not checked into as.tidem() yet,
though.) The only thing that might affect a user (well, at least a usr
who is doing unrecommended investigation of the internals of a tidem
object) is that the cos() terms are stored before the sin() terms, in
building the columns of the regression matrix. However, the form of that
matrix is not documented, and so any user that is looking that deep into
the data object is working in unrecommended territory. Sticking to the
interface, that is predict() and summary(), ought to be okay, according
to the tests.