mapPlot() and mapAxis() improvements re lonlabels
See https://github.com/dankelley/oce/issues/1625 for discussion. Since
the old arguments 'lonlabel', 'latlabel' and 'sides' were not used
previously, I felt free to alter the names of the first two, adding `s`
at the end, and also to remove the `sides` argument entirely.  With help
from @clayton33, a new UI was developed, which is similar to `axis()`
notation, and which offers the user quite a lot of control on whether
and how to compose axes.