mapContour() does in-poly with 'sf' but SLOWLY
Fine-scale testing on my home machine shows that it takes 0.1s to 0.2s
for each contour in the test that is provided in the doc, which yields a
very noticable slowdown in performance, from near-instantaneous for the
old method to maybe 10s or so with the new method.  If this is the best
that can be done, then I do not think transfering this part of the work
to 'sf' makes sense.

However, I have not looked into this carefully yet, and this commit is
mainly as a steppingstone.  What I want to check next is whether there
is a way to vectorize the sf::intersects() call.  I would be very
surprised, if not. If doing that speeds things up, then I'd be happy to
transfer this part of oce from 'sp' to 'sf'.

PS. the idea of switching this to 'sf' is to have *all* geometry and
map-projection code being done by a single library, wich ought to be
less confusing than using multiple libraries.  (Also, 'sf' seems to be
an improvement on, and eventual replacement for, 'sp'.)