Adjusting font sizes NOTE:
My formula for decreasing cex does not match that used by par(mfrow).
So, I am going to go for a new rule that if the stack has 1 or 2 panels,
use full size (cex=1) but otherwise drop to cex=0.66 as par(mfrow) does.

My old rule might have reflected a condition for e.g. an NxN grid, but
plot,adp-method() uses an Nx1 grid.

Other things I'm working on:

1. the time-range font is too big

2. I think I am doing a double-negative, in th way I'm calling these
things.  I think I ought to be doing a par(cex) before the loop, and
using cex=1 in all the function calls.

PS. this checking will fail build-tests, because a lot of functions are
spewing a lot of message() output at this moment. This is because I am
doing some rapid coding and so I even have some message("blah");stop()
blocks in place. This branch is *not* for users, only for me.