mapPlot(): cleaner debugging output (re speed tests)
Also, now the optimize() call gets a tol argument, which sets the
precision to 1/2000 of the y span of the plot. That ought to be enough
to ensure visually-correct tick marks. That only gives us a factor of 2
or so increase in speed, which is helpful but not dramatic.

Below is the expected speedup, namely 11 map-projection calls, as
opposed to 60.  That ought to be a factor of 5 or so, but it seems to
work out to be a factor of 2 or so, in my tests. Maybe there is other
overhead that I've not considered.

```R
> log2(180/.Machine$double.eps)
[1] 59.49185
> log2(2000)
[1] 10.96578
```