attempt to recover from bad ad2cp checksum
This is just a steppingstone commit, for I still have not found a
solution to the problem I see after a checksum error. I tried adding
code to search for next 0xA5 but that could come up randomly, with 1/256
probability.  So, my next step will be to test multiple bytes. Byte 2
should correspond to either 10 or 12 decimal, for example (at least in
the present file format) and byte 3 will perhaps be constant for a given
file, so reduces the probability of erronous hits.

However, I need to understand just what has gone wrong with this file,
in the bad-checksum area.  For example, are there missing bytes, or
extra bytes?

Maybe my solution should be to skip back to the last good sync+data, and
then to search forward from there, trying to get a header.  (In other
words, maybe I should not even trust the previous data-length
information.)