speed up reading of certain ODF files
Some files have thousands of 'PROCESS' header lines -- in one file
provided to me, it was over 10,000 lines -- and reading them was very,
slow.  (I didn't even bother waiting for it to complete.  Maybe it would
minutes, maybe hours.)

The problem was in renaming the 'PROCESS' lines so they could be saved
in a list within the object's metadata.  I was doing that with a loop,
where I cycled through the integers until I found one that wasn't used,
and then tacked that integer into the variable name.  This is fine if
you have a half-dozen variable-name duplicates, which I think is the
common case, since for n lines, the processing time will be proportional
to n^2.  With n=6, no problem.  With n=10000, well, problem.

I switched to using a list of counters.  Now, reading these files takes
under 5 seconds.

"Making a list, and checking it twice" is good for more than Santa.