improve some XY and XYSF comparisons
This is because the results can differ by small amounts, e.g. I changed
one all.equal() to have tolerance=1e-3, because surely a 1mm difference
is not something to be concerned about.  Also, we now check only when we
can, because sometimes either the rgdal or the sf call can make Inf or
NA, and I'm not sure they always match.