plotTS() has improved auto-lim calculation
See https://github.com/dankelley/oce/issues/1730. The problem occurs
e.g.  in calculating Slim, if either extremal S value happens to
correspond to a T value that is NA. (This can happen with data that have
created by handleFlags(), because S and T flags are independent.)

Since the goal is to auto-scale a graph -- meaning to the data that will
be displayable on that graph -- I changed the behaviour to start by
trimming any data that are not finite in both S and T.

PS. when I say "S" I mean either Practical or Absolute Salinity, and
when I say "T" I mean either Potential or Conservative Temperature. The
choice is set by the `eos` argument.