move tide test to testthat; test cm things for covr
Also, the old tide test was also used for generating some datafiles, so
I put that into create_data, although without a Makefile, so we do not
accidentally trigger an unneeded rebuild. No rebuild is needed so long
as R can read a gzipped table and do simple arithmetic and trigonometric
computations.  Rebuilding would destroy git history, for no good reason.
And there's a bit of a risk, for someone might alter this tricky file,
not knowing that it was used many years ago for a low-level setup.