plot() works with aquadopp (1773)
See https://github.com/dankelley/oce/issues/1773

I need to see if this still works for other types of adp, but I'm
guessing it will be okay.  Basically, I just check to see if the
subsetting thing is a matrix or not, and tweaked the max(abs()) call
accordingly to set a scale.