plot(ctd,eos="gsw",debug=1) now works
There was a bug in a debugging line, which did not affect calls with
debug=0, and so I do not consider this terribly important, and certainly
not enough to require a new CRAN release (which would be rejected, since
oce was last released only a week ago).  Still, it's worth fixing, and
it was just a matter of moving a line.  (I also improved the text of a
couple of other debugging lines)

The problem was using a value before defining it.  Since this only comes
up when debugging is turned on, perhaps it's not surprising that no
users reported this.  I noticed it while I was working on

https://github.com/dankelley/dfo/issues/1