change some donttest blocks (issue #1798)
See https://github.com/dankelley/oce/issues/1798, which was inspired by
a problem that was found within a donttest block
(https://github.com/dankelley/oce/issues/1795) and thus indicated that
CRAN is now running code within donttest blocks.  So, we ought to use
those blocks only for slow code that we know works.  For code that won't
work, e.g. because of the assumption of data not provided with the
package, we should use dontrun.