as.section() handles argo objects better (#1791)
The problem was that it wanted to find waterDepth.  This made sense when
the list was assumed to be a ctd object, which usually has
metadata$waterDepth (whether finite or NA), but it does not make sense
if it is a list of argo objects.

See https://github.com/dankelley/oce/issues/1797 for details.