adapt to new testhat
Some tests such as

  accyBurstMatlab <- c(-0.3144531, ETC
  expect_equal(d1[["accelerometery", "burst"]][1:10], accyBurstMatlab, scale=1, tolerance=1e-6)

started to fail with testthat-3.0.3, and the pattern seems to be that
the last digit is off by a value or two.  But I also notice that
expect_equal() seems to treat the 'scale' differently, and I don't
actually see it in the docs anymore, so I dropped that.

I went through all tests that started making errors and fixed them all
up, usually by changing a tolerance of, say 1e-6 to 3e-6 or so.  That
seems to work in pretty much all cases, I think.  But, the thing is:
these are just test of consistency, not real tests against theory.  And,
frankly, I don't even know what rounding rules were in place when I made
the tests.  I think in every case, the changes are down below realistic
limits in terms of data.

I did not have to change every test.  My impression is that this partly
has to do with, well, the *scale*.  In some cases, I just multiplied to
make the computed value and the reference value both be of order 1, and
then tests worked out to the digits I had used before.

None of this is interesting enough for me to worry about it.  I trust
the testthat authors are doing things reasonably, switching between the
two schemes appropriately ... because that's another new wrinkle.

However, I really like how testthat is now more stringent.  For example,
it won't convert POSIXct and POSIXlt things to be in the same subclass,
and that revealed an inconsistency, in which `met` subclass data were
using POSIXlt, in contrast to (I think) pretty much every other
subclass.

I think the new testthat found a few other consistencies, but I didn't
bother making detailed notes. Basically, I think it's good, and I hope
the CRAN machines won't (whenever we re-submit oce) be using different
versions that are inconsistent.  I guess we'll see.

NOTE: this "new" testthat is only 2 weeks old, but it *is* on CRAN, and
it must be one of the most-used packages on CRAN, so I don't think it's
a risk to rely on the new version, as oce now does.