isolate sf calls in hopes coverage checks work
<rantish comment>

I don't see why the action doesn't just install "sf", since I've altered
the yaml to make that happen.  But maybe this will work.  Frankly,
test-coverage is such a pain that I don't know why we bother.  I've had
a half-dozen commits in the past hour, trying to figure this out.  So
the commit history is an ugly mess of code changes (which are important
to users) and all these tweaks to actions (which say more about actions
and test machines than they do of the user experience).

</rantish comment>