handle Binary Fixed Attitude header (crudely)
The signifier is byte code 0x00 0x30.  For now, this is stored as 34 raw bytes, in metadata$binaryFixedAttitudeHeaderRaw.  You might ask -- why don't we try to read the data into subcomponents?  Well, the answer is that the docs provide no guidance on how to do that, as far as I can see.  The docs just give names of items and byte counts.  But are those things int or float? Signed or unsigned? Little or big endian?  And -- here's the kicker -- one of the items is 3 bytes, so I don't think this can be binary at all ... maybe it's BCD, or ASCII? The test files that I have available have all zeros after the 0x00 0x30, so that doesn't give me any guidance.