remove pkgdown.yaml (broken and unneeded)
It took 360 minutes and stopped, while doing some 'brew' operations.
This was unrelated to oce so there's not much I can do about it.  Also,
this had not occured in the previous git-pushes, of which there were
several that day. So the action is simply broken. I have no idea how to
debug broken actions.

Besides, it's useless.  I think at one point it actually did what I
wanted -- building up docs automatically.  But it does not seem to have
been doing that recently, as I discovered by the fact that the
representation of README had not "picked up" changes made in the past
few months. Then I noticed that it was not updating for versio 1.5.  It
was doing something, burning up CPU time somewhere, but it was *not*
updating our webpage.

Now for a third thing.  During the release to CRAN, the system asked me
if the docs directory was up-to-date.  This makes me think the
recommended new scheme may be that the author do the updating, not that
an action does it.  And that's simple enough: just use

    pkgdown::build_site()
    git add .
    git commit -m 'update webpage'
    git push

perhaps an hour prior.  I have And all I was doinSo the
problem

I don't have the patience to try googling to find