drop recent read.met test for windows
This is a new test.  It passes on 2 flavours of linux, and on macos
(both locally and remotely) but fails on windows.  So I put in a block
to just skip the test on windows.  (See comments in
tests/testthat/test_met.R)  I may end up just deleting this test, since
it seems to relate to some quirk of endianness on one platform. Not
having a windows machine makes it hard for me to find the error.  If a
user encounters problems, they will likely be able to run tests for me
on a faster turnaround time than I have with offsite checks.  (I can ask
them to print names(d@data), for example, instead of checking it against
an expected list ... and that might reveal the problem.)