change 3 DOIs to plain-text
These have been failing on winbuilder and rhub for days now.  They seem
fine with a local urlchecker::url_check(), but I think winbuilder is
more representative of CRAN, so I am going to go with that.  I have
noticed discussion on devel mailing lists about url failures, and some
people are asking that R do some deeper checking, but I doubt that they
want to slow down processing to do that.

The negative of changing from \doi{stuff} to DOI:stuff is only that a
user has to copy/paste.  The 3 links in question are journal links, so I
think they will be relatively stable (and a browser always seems to get
through) so I don't see the problem for the user ... certainly not
enough problem that we want to risk CRAN rejection for it.  I think any
user who finds a broken link is very likely to just search for the
journal-article title, and so they will get the paper in 10 seconds
instead of 1 second ... no big deal.