wrap a kriging example in dontrun
This is because it sometimes works, and sometimes fails, on
R-devel/linux CRAN machines.  And one of the questions you have to say
"yes" to asks whether all errors on the CRAN check page have been
addressed, so we'll need to fix this eventually, if this reflects
changes that are coming to those libraries, or to R itself.

Having said all that, I am a bit divided on the utility of this, and if
I see that the errors go away in a few days, I'll remove the \dontrun{}
command.