reformulate read.ctd.aml()
I had thought that all files would include a "Comments:" line, but that
is not the case.  During discussion with CR, a decision was made to try
to find the data by counting commas in the first line, and then reading
data from later lines that hold the same number of commas.  This works
for our existing test file (which holds a "Comments:" line) but it has
not been tested on other files.