skip tests for read.met()
Honestly, I just give up with this chase-the-encoding game for windows,
debian, etc.  The docs now embrace the sample code in \dontrun{}, and
the tests skip over all Windows machines.  This may be overkill because
I think many windows machines will work -- it depends on the version and
the local encoding on the machine.  I think R for Windows is changing
with respect to encoding and character topics in general.  Anyway, the
docs now say how to use read.csv() and then as.met() and therefore users
will be able to get results in maybe 2 minutes instead of 2 seconds.
Risking CRAN rejection (at submission or at a later time) is too high a
cost to pay for those 118 seconds.