remove read.met examples
They just are not worth the hassle they cause, in terms of checking on a
variety of test machines that have different OSs, different versions of
R, and different encodings.  The matrix of OS*Rversion*encoding is
fairly large, and I don't know any easy way of testing them all.  The
problem is that if any CRAN system (and I cannot replicate them
properly, so far as I can tell) has a problem, then we get a big red
'ERROR' message on the CRAN check page.  And, for what? So we can show
something that ought to be obvious to anybody reading the man page.

Note that there are also tests, and that I turn these off for Windows
machines, which seem particularly prone to revealing problems.