improve plot,argo-method re col arg
I was doing some if(missing(col)) silliness, instead of just
straightforwardingly making col be an arg to the function.  I put this
down to not really "getting" ... args, way back 15 years ago or whenever
I wrote this code.  Also, since I'm in "pardon moi" mode, I view this
plot method as a crude tool.  Any user who is familiar with argo data is
likely to work differently, e.g. using as.ctd() on individual profiles,
or maybe using as.section() for a bunch, etc. The advantage of convering
to other data types is that then a lot of plot code becomes available.