tentative change to pressure reading for ad2cp
The problem is that the docs call it 32-bit unsigned, but R does not
allow that type.  But this is stored in 0.001dbar units, and

2^31*0.001

is 2147484 (i.e. 2147 km), which is WILDLY in excess of any value, so I
think there is no chance that we can get large enough values to cause a
problem.