use v2 for pandoc and setup-r
This is something I saw in the R-devel mailing list.  I don't have a
problem with the older versions (I think) but it seems sensible to
upgrade, if new things are available.

Of course, if I see that the actions failed when I check again in an
hour, I'll revert.