ad2cp test 2: use pressure O(270), not O(7)
A recent recoding reads pressure as 4 bytes, insetad of 2 bytes.  So,
the old consistency check against 7-ish dbar had to be changed to a
check against 269-ish dbar.

NOTE that is is only a consistency check. It is not a check against
values inferred from nortek software ... something we might get in the
future.