small diff (maybe wrong) for depth inference
I'm using an incremental patch for the first time here, so perhaps I am
accidentally committing more than I want to.  But, heck, it's the
weekend and I'm going to assume nobody else will get clobbered by this,
if it happens to take me an hour to do a fuller job.