steppingad2cp steppingstone update for #1954
It now reads the built-in-tests properly again, so I am back to making
it work on Clark Richard's two test files.  It seems to work fine on the
first one, but not on the second.  That second reports checksum errors,
which I am not accustomed to seeing, and so something is wrong.

NEXT: before worrying about Clark's files, I will make it verify the
header checksum (in addition to the already-verified) data checksum.
Once I get that working on the previously-supplied test files, I will
return to Clark's files.