read.adp.nortek.ad2cp() returns adp object again
The *deep* level work is showing signs of success, so I got rid of the
premature return of a partial object.  A check tells me that, for the CR
test file in issue 1954, we are reading the raw altimeter data just as
before.  I am still saving it with "NEW" in the name, though, although
I'll delete that in the next pass.

Next steps: start doing this for all fields that are present in my test
file.  Then do it for other fields.  (All of this within
burstRawAltimeter, though ... the other items can come later. I need to
get the code in clean form first, before copy/modify work starts.)