use configuration[,i] directly (for 'burst')
See https://github.com/dankelley/oce/issues/1980#issuecomment-1186503579

I need this because the meanings of the columns vary with ID type, e.g.
bottomTrack is different from burst/average.  I had originally coded
this with burst/average in mind, and had not bothered to check through
the rest of Nortek (2017) to see if the meanings varied with ID type.

I need to settle this for the vectorization of reading.

**Next Steps.**

1. Handle average, which is a simple copy/paste exercise.

2. Work through the others that have had vector-read coded.

3. Do this in the addition of new vector-read code.

Obviously, tests will have to be run on each change.