fix recent i0v bugs
I need to remember that my commandline R is R-devel, whereas the R that
RStudio uses is R-release.  So, when I build within one system, the
package is stored *only* in the directory for *that* version.  This
explains some mysteries I've been seeing, in which changes don't do what
I expect.  Silly me.