finish vectorizing ad2cp (but...
... I am going to close issue

https://github.com/dankelley/oce/issues/1980

but that doesn't mean that I know that the code works.  We have some
test values for the 'average' and 'burst' ID types, but not (if I recall
correctly) any other type.  For other ID types, we have very little in
the way of corroboration, and I wonder about some values, which seem odd
to me.  See also

* https://github.com/dankelley/oce/issues/1979
* https://github.com/dankelley/oce/issues/1982
* https://github.com/dankelley/oce/issues/1983