fix nortek checksum problem
I was delighted to get some code from Nortek, so I could solve the
guessing-game that has frustrated me for years, as definitions in
various Nortek documents shifted over time. In this new version, the
checksum works on all the test files that I have available:

1. The 3 files in my private 'tests' suite, i.e. that are run when
   build-testing oce from github source.

2. Several other files that are stored in
   'tests/testthat/local_data/ad2cp', but not used in the test suite.

The only sample file that gave problems before was an "averaged" one
that @richardsc had provided to me. And, lo and behold, that now works
(see below), as do all the other test files.  I am closing this issue.

```
    ## Loading required package: gsw
    files <- list.files(".", ".ad2cp") for (file in files) {
        cat(file, "\n")
        print(read.oce(file, which="?"))
        cat("\n\n")
    }
    ## ad2cp_01.ad2cp
    ## warning : ldc_ad2cp_in_file() EOF before end of chunk 13 at
    cindex=42802
    ##        0x16=average    0x1c=echosounder 0x23=echosounderRaw
    0xa0=text
    ##                   7                   2                   2
    1
    ##
    ##
    ## S102791A002_Barrow_v2_avgd.ad2cp
    ## 0x16=average    0xa0=text
    ##            6            1
    ##
    ##
    ## S102791A002_Barrow_v2.ad2cp
    ##             0x15=burst
    ##                    648
    ## 0x1a=burstAltimeterRaw
    ##                    108                      1
```