give up on codecov (issue 2004)
Life is too short for this silliness. Codecov seemed like a neat idea
some years ago, but it causes lots of problems, for very little benefit.

I recall similar problems that were solved by doing some kind of setup
on either the codecov site or the github site (the former needs to be
able to access the latter ... but surely the fact that the repo is
public is enough for that?!?!).

A developer can always run codecov locally to check on things.

So long, codecov.  Good luck in your journeys.