chop out non-moving parts.
NOTE: we are only passing the data slot, so it is tricky to get time.
The code could be changed to handle if we pass the whole object, at
which point time might work more often because of how [["time"]] is
handled. But not all data will have this, I think, and so my present
scheme might be good.