Update stale.yml
I changed this from 100 days to 30 days because a month seems like a more reasonable estimate of how long a developer can (or will) keep the issue in their thoughts.

Keeping issues alive when they are no longer being actively considered has a negative consequence: active issues can get buried in a list of inactive ones.  Since the oce developers use issues as a sort of to-do list, this can have the effect of hampering development as a whole.

Note that github has a "pin" mechanism to prevent issues from disappearing. This is very handy for things that are still on the developers' to-do list, but that are complex or of relatively low priority.

Staleness is for issues that are basically moot, or that have been resolved already, but just not closed.

I think 30 days to go stale is right.  A week is a choice made in some projects, but that is too short in an academic environment, where it's often not easy to find an hour spare for a few weeks.