These changes are for issue https://github.com/dankelley/oce/issues/2024, and are being pushed to the `rsk` branch.  Things will be merged into the `develop` branch when things seem OK.  This may be a few weeks, given end-of-term rush and the presence of 3 persons on the development stream.
The new features are as follows.  (Well, perhaps there are others; see
the link given above, and -- depending on who you are -- a parallel
private email thread for more.)

1. Using e.g. `read.rsk(filename, debug=1)` now reports the names of all
   the tables in the RSK file.

2. If there is a table named `geodata` then it is read, and stored in the
`metadata` slot of the value returned by `read.rsk()`.

3. Since the test file I'm working with has a different number of `geodata`
samples than actual CTD samples, I have not coded a way to synch these things
up.  I think the timestamp is the way to go for that, although I would not be
surprised to get weird problems, e.g. perhaps UTC used for one system and a
local time used for another ... I'd like to encourage Clark and the
as-yet-anonymous issue generator to weigh in on this, and to try some tests.

4. In addition to the previous point, I'd appreciate Clark's advice on other
data tables that we might want to incorporate.  I suppose we could just read
every single one of them and shove them into the `metadata` slot, so that a
user could do whatever they want. That might yield large objects, though, and
that's a concern for many users who might not have large-memory devices.  I
haven't even examined the issue of size, though ... I'm just talking about
general issues that tend to come up in oce, where the preference is often to
start with things that are of known importance (like pressure, etc. for
CTD-like data).