decrease stale time
The problem with longer times is that it takes quite a lot of time to "spin up" on old things that are no longer top of mind.  That goes for developers and reporters alike.  It's easy to pin issues to keep them from going stale, so I see no negative aspects to decreasing the stale time, at least during off-teaching months when I have more time.