tidem() returns frequencies in order
Need this sometimes, e.g. if the user supplies constituents that not in
frequency-specific order.  Also, I think this may happen because of
constituent substitutions (but I've not tested that).  Frankly, all I
really know is when I was thinking about how to work with 6-hourly data,
some of my tests gave results that were not ordered by freqency.  Since
the default data(sealevel) case *is* ordered in that way, this change
will not affect the test suite.  I can't see what harm it would do to
sort something that is already sorted, and I do think that users might
expect results to always be sorted by frequency, because that seems like
a sensible (probably standard) representation.