applyMagneticDeclination for cm,adp,adv
This is done with a new generic function.  I suppose that is overkill,
but it is a bit more DRY and doing it this way exemplifies a path for
future things of a similar ilk.

I have a graphical test in my sandbox but need to add some actual tests
to the test suite before moving on to my next task: handling
'declination' in various adp and adv functions.