convert ## comments to # comments
These remnants do not match the current style convention.  It is hard to
believe I could have made an error with this, but, just in case, I did:

1. a full local build, test and then check (using RStudio and also a
   Makefile)
2. an external (win-builder) check with R version 4.2.2 (2022-10-31
   ucrt)

and saw no issues with either on.  Given this, I am not going to bother
letting it sit on my machine for days just in case errors appear.  I'm
committing right away, in case CR wants to pull to get the latest ad2cp
code.