revert some tests of the odd RDI byte sequence
The tests (now reverted) of the odd byte sequence convinced me that
we'll need some RDI documentation on this format to make any progress.
There's just no point in guessing on things, e.g. "maybe it's not the
10th byte that tells us such-and-such but the 11th" because every test
that involves changing C code means a slowish rebuild process.  This is
different from a change-resource cycle.

The changes as committed here are an improvement to "develop" in the
sense of having slightly more informative messages, and so I am
committing this with the intention of merging in a few moments.