move plot tests to local-only (issue 2073)
A CRAN member used the words 'obfuscated' and 'unhelpful' in describing
the oce test suite.  This description was in an email referring to what
seemed to be an error with oce, but that later turned out to be an error
with how the test machines were set up (I think).

In any case, I decided, in consultation with CR, to move some of the
tests of plotting in such a way as to prevent their incorporation in the
source tarball.  Perhaps this might make the tests less 'obfuscated' and
less 'unhelpful'.  And it should shorten the testing time, although I
can't imagine it will let oce processing complete in under the 10-minute
threshold that threatens CRAN rejection.