fix encodings (remove non-ascii symbols)
The bad symbols include
* Greek lower-case mu
* slanted " marks
* long dashes
* degree signs

I am not 100% sure I need to do this, but some files were not building in oce, since I updated macOS from one Sonoma beta version to another, today.  I don't see the harm in using ascii instead of utf, anyway.  There is one exception: a table in read.sbe.R holds a theta character, but it is marked with the recommended latex magic, so I think it will be okay.