perhaps solve CR aquadopp problem
We don't have an issue on this -- communication was carried out over email.  The problem was a file that had an unreadable serial number.  Upon examination, I saw that the string contained a 0x80 character, which cannot be converted to a character (at least with the encoding on my computer).  As a test, controlled by a logical value set at the top of R/adp.nortek.R, we now examine the 14 bytes that hold the serial number, and search for *any* character outside the range 0x00 to 0x80.  If we find such a character, we delete it an all following characters in this 14-byte chunk.  I checked the C code we got from Nortek quite a while back, and confirmed that the serial number is supposed to be 14 bytes.  Whether other things are read correctly is to be determined, and discussed either over private email or within a github issue.