fix flags vignette (was cut off)
It cut off at a single-quote mark.  I checked it in an editor, and it was a normal single-quote mark as made on my keyboard (and as occurs in other spots of the document).  Still, I reworded that a bit.  I also went through the paragraphs, reformatting them to an approximately constant width.  That was actually more so I could run my eyes over the text to see if my editor reported problems (like non-printable characters).